completed/05_AI_기업_블로그_수집_에이전트/requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.3
requests>=2.31.0
